Prototype Concept for Channel Properties API
Working on a callback system so that the lifetime of the channel is
limited to the lifetime of the call.

Not sure if it is better than a channel struct that borrows the task for
it's lifetime.

All does allow checks for a valid channel and type first though. Does
this matter?